Move cat types into their own package
