--- src/main/kotlin/dev/akif/cats/CreateCat.kt
@@ -1,5 +1,0 @@
-package dev.akif.cats
-
-import dev.akif.crud.CRUDCreateModel
-
-data class CreateCat(val name: String, val breed: String, val age: Int) : CRUDCreateModel
