--- src/main/kotlin/dev/akif/cats/CreateCatDTO.kt
@@ -1,5 +1,0 @@
-package dev.akif.cats
-
-import dev.akif.crud.CRUDCreateDTO
-
-data class CreateCatDTO(val name: String, val breed: String, val age: Int) : CRUDCreateDTO
