--- src/main/kotlin/dev/akif/cats/UpdateCat.kt
@@ -1,5 +1,0 @@
-package dev.akif.cats
-
-import dev.akif.crud.CRUDUpdateModel
-
-data class UpdateCat(val name: String, val age: Int) : CRUDUpdateModel
