--- src/main/kotlin/dev/akif/cats/UpdateCatDTO.kt
@@ -1,5 +1,0 @@
-package dev.akif.cats
-
-import dev.akif.crud.CRUDUpdateDTO
-
-data class UpdateCatDTO(val name: String, val age: Int) : CRUDUpdateDTO
