+++ src/main/kotlin/dev/akif/cats/cat/CatDTO.kt
@@ -1,6 +1,6 @@
-package dev.akif.cats
+package dev.akif.cats.cat
 
-import dev.akif.crud.CRUDDTO
+import dev.akif.crud.*
 import java.time.Instant
 import java.util.UUID
 
@@ -16,3 +16,7 @@ data class CatDTO(
     override fun createdAt(): Instant = createdAt
     override fun updatedAt(): Instant = updatedAt
 }
+
+data class CreateCatDTO(val name: String, val breed: String, val age: Int) : CRUDCreateDTO
+
+data class UpdateCatDTO(val name: String, val age: Int) : CRUDUpdateDTO
