+++ src/main/kotlin/dev/akif/cats/cat/CatMapper.kt
@@ -1,6 +1,6 @@
-package dev.akif.cats
+package dev.akif.cats.cat
 
-import dev.akif.crud.CRUDMapper
+import dev.akif.crud.*
 import org.springframework.stereotype.Component
 import java.time.Instant
 import java.util.UUID
@@ -37,4 +37,30 @@ class CatMapper : CRUDMapper<UUID, CatEntity, Cat, CreateCat, UpdateCat> {
             age = updateModel.age
         }
     }
+}
+
+@Component
+class CatDTOMapper : CRUDDTOMapper<UUID, Cat, CatDTO, CreateCat, UpdateCat, CreateCatDTO, UpdateCatDTO> {
+    override fun createDTOToCreateModel(createDTO: CreateCatDTO): CreateCat =
+        CreateCat(
+            name = createDTO.name,
+            breed = createDTO.breed,
+            age = createDTO.age
+        )
+
+    override fun modelToDTO(model: Cat): CatDTO =
+        CatDTO(
+            id = model.id,
+            name = model.name,
+            breed = model.breed,
+            age = model.age,
+            createdAt = model.createdAt,
+            updatedAt = model.updatedAt
+        )
+
+    override fun updateDTOToUpdateModel(updateDTO: UpdateCatDTO): UpdateCat =
+        UpdateCat(
+            name = updateDTO.name,
+            age = updateDTO.age
+        )
 }
